Make the RadioThread keep multiple queues, indexed by integer priority, 0 is highest.  Modify the submit method to take a priority argument, assuming highest.
In MemoryEditor, use this in the following ways:

In MemoryEditor::__init__(), start a background load of all the
radio's memories (prio 10) so that they can be loaded from cache instead
of fetched from the radio.  Start the background load at the upper limit
of the UI control, which will cause us to only load the ones higher than
their preference setting for initial loading.

In MemoryEditor::prefill() load memories at prio 2. All others (such as
a write following a user's modification to a channel) are sent at prio
0.

This means that the memories loading in the background take absolute
lowest priority.  While loading memories for the main display, an edit
gets sent to the radio immediately.  Switching to the D-STAR tab loads
the callsigns at prio 0 since the user clearly just asked for them.