Allow cancelling the clone thread(s)
This adds a "KillableThread" class to chirp_common which has a kill()
method that takes an Exception and interrupts the running thread with
it.

By changing the CloneThread class to inherit from KillableThread and
adding a button to CloneProg, we can allow the user to cancel a clone
mid-flight and avoid calling the callback function to terminate the
operation.