Add TH-D72 radio support
The TH-D72 radio is a clone radio similar to the TMV71
but it had enough differences that it merited a separate
driver.  In addition to those differences, I also used
the new bitwise parser to deal with the raw memory data.

I have tested adding, modifying, deleting, upload,
download, writing to file, and reading from file.

Signed-off-by: Vernon Mauery <vernon@mauery.org>