Add power level infrastructure
This allows radios to declare symbolic power level names and the associated
absolute power level.  This enables adjustment of the level in the UI,
but also allows import to attempt to choose a level on the destination
radio that closely matches that of the source.

By using the new PowerLevel object, radios can attach an alternate power
level to a memory to indicate that a given band or other condition causes
the power output to be different than one of the normal options (i.e. 220MHz
on a vx7).  The UI deals with the power levels only at the symbolic level,
and thus will attempt to set a power level equal to what the user chose.
Since we don't expose the absolute power level to the user, the VX7 can
handle the conversion internally, and import will, like other cases,
make it graceful as well.