[uv5r] Remove the code attempting to identify the radio version
When the 291 firmware came out, it looked like we were going to have
to identify the version from the string that gets sent back during the
initial start-clone-mode handshake. It was never reliable, and apparently
not very identifying anyway. Since we now have the ability to identify
the actual firmware version from the aux block, this change makes the
identification routine just send known magics until the radio responds.

Should fix #373