[anytone] Fix some issues in the indexes and model detection
The previous patch didn't properly rename the Powerwerx class, and
it was also getting detected incorrectly. This expands the header
data that gets examined for the identifier and makes the base anytone
more specific.

Also, the powerwerx settings block looks different, so disable settings
for the moment.

Related to #1031 and #1365