[developer] Add diffs-only mode to file diff utility - #1699
Adds an alternate display mode to the developer tools whole-file hex
dump/diff utility in which only lines which are different are printed.
Ranges of one or more lines that do not differ are replaced with a
single blank line.

This mode is invoked by setting the first memory location in the
"Diff Radios" popup to -2. The existing functionality invoked by -1
is not affected.

#1699