[KGUV8D] Decode oem memory as ascii string
Address issue where the UV8D driver was not decoding strings from memory,
but rather was treating them as raw memory which was preventing the app
from displaying the raw memory as a string (which it can not do in cases
where the string contains null characters, etc..);

This bug was causing the settings UI to not display any settings due to
the exception being thrown by the radio driver;

It was not caught by the unit tests due to the fact the img checked into the
tree for this radio is of a diff. version than the image reported by the user
in the bug report.

Bug #2377