Use logging in run_tests (#2347)
We need to use logging in run_tests, if only to catch the new output
being generated by the drivers.  This patch captures a debug.log for the
entire test run while squelching output on the console.  The prevents
spurious error messages from being displayed.  This may be improved upon
in subsequent patches, but this is necessary to prevent anyone from
seeing confusing output when running the tests.