[AP510] Disable AP510 driver. In response to #3617
Only a buggy, outdated firmware version is supported by Chirp. Newer firmware
revisions changed the programming protocol and are incompatible. To avoid
confusion and discourage running buggy firmware, this patch disables the AP510
driver in Chirp.

The code should remain, as it is a great starting point for anyone wishing to
write support for the latest AP510 firmware.