[PATCH][KT8900] Comment about old versions of KT8900 in the UI, fixes #3539
All points that the QYT KT8900 is the main OEM model, and there are two
variants of it, from the prespective of the communication dance:

* One is with just the ID and right to the down/up process (clone of WACCOM
Mini8900)
* The other (is the default now) with a second id read & a extra step on the
upload process.

As both uses different communications steps we can't put them both under the
same case, fortunately the "older" model has a clone in the WACCOM Mini8900.

We put here a notice on the failure to read a KT8900 to warn the users to try
the WACCOM Mini8900 instead.