[NC-630A] Driver Cleanup
The main purpose of this patch is a general cleanup of the kyd.py driver. Some
things were tweaked to more closely match how the "factory" software behaves
during the cloning process. Also, code was also added to be more selective in
detecting KYD NC-630A images.

Except for adding an alias for the Plant-Tours MT-700, this patch does not add
any new features and does not make any changes that will be apparent to the
CHIRP user.

Related to #4235